refactor(chat): extract session tree parser and prune dead fork internals
