+++ CHANGELOG.md
@@ -36,6 +36,7 @@ The format is inspired by [Keep a Changelog](https://keepachangelog.com/en/1.1.0
 - Added shared composer-textarea sync helpers in `chat-view` and reused them across input staging/history/preview/clear flows; also extracted preferred model-picker provider resolution into `src/models/model-selection.ts` and centralized session transient-state reset helpers for project/session switches.
 - Split composer UI rendering into dedicated view modules: `src/components/chat-view/composer-controls-view.ts` (model picker + send/thinking controls) and `src/components/chat-view/composer-fragments-view.ts` (queued pills, attachments, skill draft pill), reducing `chat-view` template sprawl and tightening component boundaries.
 - Further split composer rendering into `src/components/chat-view/composer-slash-palette-view.ts` and `src/components/chat-view/composer-stats-view.ts`, and extracted keyboard/event orchestration to `src/components/chat-view/composer-input-events.ts` (with thin `chat-view` delegation), reducing inline template logic/coupling and keeping composer behavior modular.
+- Extracted session tree parsing + tree-line prefix logic from `chat-view` into `src/components/chat-view/history-tree-utils.ts` (including role/entry mapping + active-path derivation) and removed stale fork timeline expansion internals that were no longer wired into rendering.
 - Compaction status rendering was reduced to a minimal workflow-style row with collapsed-by-default details instead of a heavy status card.
 - Auto-rename extension recommendation and desktop config bridge now target `@byteowlz/pi-auto-rename`, with dynamic command-to-package resolution for config-intent slash commands (including `/auto-rename config`) instead of hardcoded package-name routing.
 - Recommended notifications extension now defaults to `pi-smart-voice-notify`, and Packages auto-migrates legacy `pi-desktop-notify` installs by installing the new package and removing the old one.
