+++ src/components/chat-view.ts
@@ -55,6 +55,7 @@ import {
 } from "./chat-view/composer-input-events.js";
 import { renderSlashPaletteView } from "./chat-view/composer-slash-palette-view.js";
 import { renderComposerStatsView } from "./chat-view/composer-stats-view.js";
+import { compactTreeLinePrefix, parseSessionTreeRows } from "./chat-view/history-tree-utils.js";
 import {
 	displayProviderLabel as displayProviderLabelFromCatalog,
 	isOAuthProviderId as isOAuthProviderIdInCatalog,
@@ -130,25 +131,6 @@ interface Notice {
 interface ForkOption {
 	entryId: string;
 	text: string;
-}
-
-interface ForkTimelineRow {
-	main: UiMessage;
-	sourceIndex: number;
-	thinkingSnippets: string[];
-	tools: ToolCallBlock[];
-}
-
-interface SessionTreeEntryRecord {
-	id: string;
-	parentId: string | null;
-	type: string;
-	index: number;
-	role: UiRole;
-	entryLabel: string;
-	preview: string;
-	displayText: string;
-	canFork: boolean;
 }
 
 interface HistoryTreeRow {
@@ -494,8 +476,6 @@ export class ChatView {
 	private historyTreeRequestSeq = 0;
 	private forkEntryIdByMessageId = new Map<string, string>();
 	private forkTargetsRequestSeq = 0;
-	private forkExpandedMessageRows = new Set<string>();
-	private forkExpandedToolRows = new Set<string>();
 	private historyQuery = "";
 	private historyRoleFilter: UiRole | "all" = "all";
 	private autoFollowChat = true;
@@ -4737,8 +4717,6 @@ export class ChatView {
 		this.historyTreeRows = [];
 		this.historyQuery = options?.query?.trim() ?? "";
 		this.historyRoleFilter = "all";
-		this.forkExpandedMessageRows.clear();
-		this.forkExpandedToolRows.clear();
 		this.render();
 		void this.loadSessionTreeForHistory();
 	}
@@ -4751,8 +4729,6 @@ export class ChatView {
 		this.historyQuery = options?.query?.trim() ?? "";
 		this.historyRoleFilter = "all";
 		this.forkOptions = [];
-		this.forkExpandedMessageRows.clear();
-		this.forkExpandedToolRows.clear();
 		this.render();
 		if (!this.historyViewerLoading) {
 			void this.loadForkTargetsForHistory();
@@ -4770,8 +4746,6 @@ export class ChatView {
 		this.historyRoleFilter = "all";
 		this.forkOptions = [];
 		this.forkEntryIdByMessageId.clear();
-		this.forkExpandedMessageRows.clear();
-		this.forkExpandedToolRows.clear();
 		this.render();
 	}
 
@@ -6311,82 +6285,6 @@ export class ChatView {
 		`;
 	}
 
-	private buildForkTimelineRows(source: UiMessage[]): ForkTimelineRow[] {
-		const rows: ForkTimelineRow[] = [];
-		let latestAssistantRow: ForkTimelineRow | null = null;
-		for (let i = 0; i < source.length; i++) {
-			const msg = source[i];
-			if (!msg) continue;
-			if (msg.role === "user") {
-				rows.push({ main: msg, sourceIndex: i, thinkingSnippets: [], tools: [] });
-				latestAssistantRow = null;
-				continue;
-			}
-			if (msg.role !== "assistant") continue;
-
-			const text = msg.text.trim();
-			const thinking = (msg.thinking ?? "").trim();
-			const tools = msg.toolCalls ?? [];
-			if (text) {
-				const row: ForkTimelineRow = {
-					main: msg,
-					sourceIndex: i,
-					thinkingSnippets: thinking ? [thinking] : [],
-					tools: [...tools],
-				};
-				rows.push(row);
-				latestAssistantRow = row;
-				continue;
-			}
-
-			if (!thinking && tools.length === 0) continue;
-			if (!latestAssistantRow) {
-				latestAssistantRow = {
-					main: msg,
-					sourceIndex: i,
-					thinkingSnippets: [],
-					tools: [],
-				};
-				rows.push(latestAssistantRow);
-			}
-			if (thinking) latestAssistantRow.thinkingSnippets.push(thinking);
-			if (tools.length > 0) latestAssistantRow.tools.push(...tools);
-		}
-		return rows;
-	}
-
-	private forkRowKey(row: ForkTimelineRow): string {
-		const base = row.main.sessionEntryId ?? row.main.id;
-		return `${base}:${row.sourceIndex}`;
-	}
-
-	private toggleForkMessageExpanded(rowKey: string): void {
-		if (this.forkExpandedMessageRows.has(rowKey)) {
-			this.forkExpandedMessageRows.delete(rowKey);
-		} else {
-			this.forkExpandedMessageRows.add(rowKey);
-		}
-		this.render();
-	}
-
-	private toggleForkToolsExpanded(rowKey: string): void {
-		if (this.forkExpandedToolRows.has(rowKey)) {
-			this.forkExpandedToolRows.delete(rowKey);
-		} else {
-			this.forkExpandedToolRows.add(rowKey);
-		}
-		this.render();
-	}
-
-	private forkRowPreview(row: ForkTimelineRow): string {
-		const normalized = this.messagePreview(row.main).replace(/\s+/g, " ").trim();
-		if (normalized && normalized !== "(empty message)") return normalized;
-		if (row.main.role === "assistant" && (row.thinkingSnippets.length > 0 || row.tools.length > 0)) {
-			return "assistant activity";
-		}
-		return normalized || "(empty message)";
-	}
-
 	private resolveForkEntryId(messages: UiMessage[], index: number): string | null {
 		const current = messages[index];
 		if (!current) return null;
@@ -6400,302 +6298,6 @@ export class ChatView {
 			if (entryId) return entryId;
 		}
 		return null;
-	}
-
-	private roleFromSessionEntry(roleRaw: string): UiRole {
-		const normalized = roleRaw.trim().toLowerCase();
-		if (normalized === "user") return "user";
-		if (normalized === "assistant") return "assistant";
-		if (normalized === "custom" || normalized === "custom_message") return "custom";
-		return "system";
-	}
-
-	private mapSessionTreeEntry(
-		record: Record<string, unknown>,
-		index: number,
-		labelsByTargetId: Map<string, string>,
-	): SessionTreeEntryRecord | null {
-		const type = typeof record.type === "string" ? record.type.trim() : "";
-		if (!type || type === "session" || type === "label") return null;
-
-		const id = typeof record.id === "string" ? record.id.trim() : "";
-		if (!id) return null;
-
-		const parentRaw = record.parentId;
-		const parentId = typeof parentRaw === "string" && parentRaw.trim().length > 0 ? parentRaw.trim() : null;
-		let role: UiRole = "system";
-		let entryLabel = type.replace(/_/g, " ");
-		let preview = "";
-		let displayText = "";
-		let canFork = false;
-
-		switch (type) {
-			case "message": {
-				const message = record.message;
-				const messageRecord = message && typeof message === "object" ? (message as Record<string, unknown>) : null;
-				const messageRoleRaw = typeof messageRecord?.role === "string" ? messageRecord.role.trim() : "system";
-				const messageRole = messageRoleRaw.toLowerCase();
-
-				if (messageRole === "user") {
-					role = "user";
-					entryLabel = "user";
-					preview = this.extractText(messageRecord?.content).replace(/\s+/g, " ").trim() || "(empty message)";
-					displayText = `user: ${preview}`;
-					canFork = true;
-					break;
-				}
-
-				if (messageRole === "assistant") {
-					role = "assistant";
-					entryLabel = "assistant";
-					const contentPreview = this.extractText(messageRecord?.content).replace(/\s+/g, " ").trim();
-					const stopReason = pickString(messageRecord ?? {}, ["stopReason", "stop_reason"]) ?? "";
-					const errorMessage = pickString(messageRecord ?? {}, ["errorMessage", "error_message"]) ?? "";
-					preview = contentPreview || (stopReason === "aborted" ? "(aborted)" : errorMessage || "(no content)");
-					displayText = `assistant: ${preview}`;
-					break;
-				}
-
-				if (messageRole === "toolresult") {
-					role = "system";
-					entryLabel = "tool";
-					const toolName = pickString(messageRecord ?? {}, ["toolName", "tool_name"]) ?? "tool";
-					const toolOutputRaw = this.extractToolOutput(messageRecord?.content ?? messageRecord?.result ?? messageRecord ?? {});
-					preview = toolOutputRaw.replace(/\s+/g, " ").trim() || "(no output)";
-					displayText = `[${toolName}: ${truncate(preview, 120)}]`;
-					break;
-				}
-
-				if (messageRole === "bashexecution") {
-					role = "system";
-					entryLabel = "bash";
-					const command = pickString(messageRecord ?? {}, ["command"]) ?? "bash";
-					preview = command;
-					displayText = `[bash: ${truncate(command.replace(/\s+/g, " ").trim(), 120)}]`;
-					break;
-				}
-
-				role = this.roleFromSessionEntry(messageRoleRaw);
-				entryLabel = messageRoleRaw || "message";
-				preview = this.extractText(messageRecord?.content).replace(/\s+/g, " ").trim() || `(${entryLabel})`;
-				displayText = `[${entryLabel}]: ${preview}`;
-				break;
-			}
-			case "custom_message": {
-				role = "custom";
-				const customType = pickString(record, ["customType", "custom_type"]) ?? "custom";
-				entryLabel = customType;
-				preview = this.extractText(record.content).replace(/\s+/g, " ").trim() || "(empty)";
-				displayText = `[${customType}]: ${preview}`;
-				break;
-			}
-			case "branch_summary": {
-				role = "system";
-				entryLabel = "branch summary";
-				preview = (pickString(record, ["summary"]) ?? this.extractText(record.content)).replace(/\s+/g, " ").trim() || "(empty)";
-				displayText = `[branch summary]: ${truncate(preview, 180)}`;
-				break;
-			}
-			case "compaction": {
-				role = "system";
-				entryLabel = "compaction";
-				const tokensBefore = pickNumber(record, ["tokensBefore", "tokens_before"]);
-				preview = (pickString(record, ["summary"]) ?? "compaction entry").replace(/\s+/g, " ").trim();
-				const tokensBadge = typeof tokensBefore === "number" && Number.isFinite(tokensBefore) ? `${Math.max(1, Math.round(tokensBefore / 1000))}k tokens` : "summary";
-				displayText = `[compaction: ${tokensBadge}] ${truncate(preview, 160)}`;
-				break;
-			}
-			case "thinking_level_change": {
-				role = "system";
-				entryLabel = "thinking";
-				const level = pickString(record, ["thinkingLevel", "thinking_level"]) ?? "updated";
-				preview = level;
-				displayText = `[thinking: ${level}]`;
-				break;
-			}
-			case "model_change": {
-				role = "system";
-				entryLabel = "model";
-				const provider = pickString(record, ["provider"]) ?? "provider";
-				const modelId = pickString(record, ["modelId", "model_id"]) ?? "model";
-				preview = `${provider}/${modelId}`;
-				displayText = `[model: ${preview}]`;
-				break;
-			}
-			case "session_info": {
-				role = "system";
-				entryLabel = "title";
-				preview = pickString(record, ["name"]) ?? "(untitled)";
-				displayText = `[title: ${preview}]`;
-				break;
-			}
-			case "custom": {
-				role = "custom";
-				const customType = pickString(record, ["customType", "custom_type"]) ?? "custom";
-				entryLabel = customType;
-				preview = this.extractText(record.data).replace(/\s+/g, " ").trim() || "custom entry";
-				displayText = `[custom: ${customType}] ${truncate(preview, 140)}`;
-				break;
-			}
-			default: {
-				role = "system";
-				preview = this.extractText(record.content).replace(/\s+/g, " ").trim() || entryLabel;
-				displayText = `[${entryLabel}]: ${truncate(preview, 160)}`;
-				break;
-			}
-		}
-
-		const resolvedLabel = labelsByTargetId.get(id);
-		if (resolvedLabel) {
-			entryLabel = `${entryLabel} · ${resolvedLabel}`;
-			displayText = `[${resolvedLabel}] ${displayText}`;
-		}
-		const normalizedPreview = preview.replace(/\s+/g, " ").trim();
-		const normalizedDisplayText = displayText.replace(/\s+/g, " ").trim();
-
-		return {
-			id,
-			parentId,
-			type,
-			index,
-			role,
-			entryLabel,
-			preview: normalizedPreview || `(${entryLabel})`,
-			displayText: normalizedDisplayText || `${entryLabel}: ${normalizedPreview || "(empty)"}`,
-			canFork,
-		};
-	}
-
-	private resolveCurrentTreeLeafId(entriesById: Map<string, SessionTreeEntryRecord>, entries: SessionTreeEntryRecord[]): string | null {
-		for (let i = this.messages.length - 1; i >= 0; i--) {
-			const entryId = this.messages[i]?.sessionEntryId;
-			if (entryId && entriesById.has(entryId)) return entryId;
-		}
-		for (let i = entries.length - 1; i >= 0; i--) {
-			const entry = entries[i];
-			if (!entry) continue;
-			if (entriesById.has(entry.id)) return entry.id;
-		}
-		return null;
-	}
-
-	private compactTreeLinePrefix(prefix: string, depth: number): string {
-		const normalized = prefix ?? "";
-		if (!normalized) return "";
-		const maxVisibleDepth = 14;
-		if (depth <= maxVisibleDepth) return normalized;
-		const charsPerLevel = 3;
-		const tail = normalized.slice(Math.max(0, normalized.length - maxVisibleDepth * charsPerLevel));
-		return `… ${tail}`;
-	}
-
-	private parseSessionTreeRows(sessionContent: string): HistoryTreeRow[] {
-		const rawRecords: Array<{ record: Record<string, unknown>; index: number }> = [];
-		const lines = sessionContent.split(/\r?\n/);
-		for (const line of lines) {
-			const trimmed = line.trim();
-			if (!trimmed) continue;
-			let parsed: unknown;
-			try {
-				parsed = JSON.parse(trimmed);
-			} catch {
-				continue;
-			}
-			if (!parsed || typeof parsed !== "object") continue;
-			rawRecords.push({
-				record: parsed as Record<string, unknown>,
-				index: rawRecords.length,
-			});
-		}
-
-		if (rawRecords.length === 0) return [];
-
-		const labelsByTargetId = new Map<string, string>();
-		for (const { record } of rawRecords) {
-			if (record.type !== "label") continue;
-			const targetId = typeof record.targetId === "string" ? record.targetId.trim() : "";
-			if (!targetId) continue;
-			const label = typeof record.label === "string" ? record.label.trim() : "";
-			if (!label) {
-				labelsByTargetId.delete(targetId);
-			} else {
-				labelsByTargetId.set(targetId, label);
-			}
-		}
-
-		const entries: SessionTreeEntryRecord[] = rawRecords
-			.map(({ record, index }) => this.mapSessionTreeEntry(record, index, labelsByTargetId))
-			.filter((entry): entry is SessionTreeEntryRecord => Boolean(entry));
-
-		if (entries.length === 0) return [];
-
-		const entriesById = new Map<string, SessionTreeEntryRecord>();
-		for (const entry of entries) {
-			entriesById.set(entry.id, entry);
-		}
-
-		const childrenByParent = new Map<string, SessionTreeEntryRecord[]>();
-		const roots: SessionTreeEntryRecord[] = [];
-		for (const entry of entries) {
-			const parentId = entry.parentId && entriesById.has(entry.parentId) ? entry.parentId : null;
-			if (!parentId) {
-				roots.push(entry);
-				continue;
-			}
-			const bucket = childrenByParent.get(parentId) ?? [];
-			bucket.push(entry);
-			childrenByParent.set(parentId, bucket);
-		}
-		const byIndex = (a: SessionTreeEntryRecord, b: SessionTreeEntryRecord): number => a.index - b.index;
-		roots.sort(byIndex);
-		for (const bucket of childrenByParent.values()) {
-			bucket.sort(byIndex);
-		}
-
-		const currentLeafId = this.resolveCurrentTreeLeafId(entriesById, entries);
-		const activePath = new Set<string>();
-		let cursor = currentLeafId;
-		while (cursor && entriesById.has(cursor)) {
-			if (activePath.has(cursor)) break;
-			activePath.add(cursor);
-			const parentId = entriesById.get(cursor)?.parentId ?? null;
-			cursor = parentId && entriesById.has(parentId) ? parentId : null;
-		}
-
-		const rows: HistoryTreeRow[] = [];
-		const buildPrefix = (ancestorHasNext: boolean[], isLast: boolean, depth: number): string => {
-			const parts: string[] = ancestorHasNext.map((hasNext) => (hasNext ? "│  " : "   "));
-			if (depth > 0) {
-				parts.push(isLast ? "└─ " : "├─ ");
-			}
-			return parts.join("");
-		};
-		const visit = (entry: SessionTreeEntryRecord, depth: number, ancestorHasNext: boolean[], isLast: boolean): void => {
-			rows.push({
-				entryId: entry.id,
-				depth,
-				role: entry.role,
-				entryLabel: entry.entryLabel,
-				preview: entry.preview,
-				displayText: entry.displayText,
-				linePrefix: buildPrefix(ancestorHasNext, isLast, depth),
-				onActivePath: activePath.has(entry.id),
-				canFork: entry.canFork,
-			});
-			const children = childrenByParent.get(entry.id) ?? [];
-			const nextAncestorHasNext = [...ancestorHasNext, !isLast];
-			for (let i = 0; i < children.length; i += 1) {
-				const child = children[i];
-				if (!child) continue;
-				visit(child, depth + 1, nextAncestorHasNext, i === children.length - 1);
-			}
-		};
-		for (let i = 0; i < roots.length; i += 1) {
-			const root = roots[i];
-			if (!root) continue;
-			visit(root, 0, [], i === roots.length - 1);
-		}
-		return rows;
 	}
 
 	private async loadSessionTreeForHistory(): Promise<void> {
@@ -6715,7 +6317,15 @@ export class ChatView {
 		try {
 			const content = await rpcBridge.getSessionContent(sessionPath);
 			if (requestId !== this.historyTreeRequestSeq || this.historyViewerMode !== "browse" || !this.historyViewerOpen) return;
-			this.historyTreeRows = this.parseSessionTreeRows(content);
+			this.historyTreeRows = parseSessionTreeRows({
+				sessionContent: content,
+				currentSessionEntryIds: this.messages.map((message) => message.sessionEntryId ?? "").filter((id) => id.length > 0),
+				extractText: (value) => this.extractText(value),
+				extractToolOutput: (value) => this.extractToolOutput(value),
+				truncateText: truncate,
+				pickString,
+				pickNumber,
+			});
 		} catch (err) {
 			if (requestId !== this.historyTreeRequestSeq || this.historyViewerMode !== "browse" || !this.historyViewerOpen) return;
 			console.error("Failed to load session tree:", err);
@@ -6838,7 +6448,7 @@ export class ChatView {
 												const visibleMessageId = sessionMessageIdByEntryId.get(row.entryId) ?? null;
 												const canJump = Boolean(visibleMessageId);
 												const title = canJump ? "Jump to this entry" : "Entry is outside the active branch";
-												const compactPrefix = this.compactTreeLinePrefix(row.linePrefix, row.depth);
+												const compactPrefix = compactTreeLinePrefix(row.linePrefix, row.depth);
 												const rowText = row.displayText.trim() || row.preview || "(entry)";
 												const lineText = `${compactPrefix}${row.onActivePath ? "• " : "  "}${truncate(rowText, 320)}`;
 												const lineBody = html`<span class="history-tree-line-mono role-${row.role}">${lineText}</span>`;
